Added various debug traces to the input system for easier tracking of problems when the input isn't coming through as one would expect
Fixed issue with the input not doing the default processing when no touch data was available - this cause the snow way to draw lines to not be updated when simply drawing in one axis and should now be more responsive.

Performance improvement: Added call to the input system update from the touch event processing which will mean the input is processed immediately rather than waiting for the next update which will further improve the responsiveness of the input.