Added Scheme asset and filetype - This is an XML file that defines Node[Tree]s all based on Refs (NodeRef, TypeRef) and data. You can import this into any scheme and it will create the node from a factory and turn all the data into an Init Message sent to the node after creation. Like all assets this is then turned into a binary .Asset file. Exporting a node tree(graph or part of a graph) back to .scheme xml not done yet but should be easy to implement. Graph's have changed so they're all based on a TGraphBase type which has no knowledge of node types, but does everything on TRef's. All graphs now support class factories. All graph nodes now store their Type (TypeRef) as well as NodeRef. A primitive RuntimeTypeInfo has been implemented using the same Type TRef's. Most of the SceneGraph extra code has gone as it's functionality is the same in the base graph type. You can now create nodes on the graph and it will be added and initialised etc. (using real values or TRef's)
Added a TGui class to turn clicks on render nodes into actions. Just relays out a simple fake Action message with the specified ref's when mouse is down or up. Also detects when mouse is still held down but no longer over the render object (e.g. finger has moved OFF the node)

TMenu asset has been implemented. Most of the TMenu code (i.e. the TMenuItems) have been moved into the asset. No file import/export yet, but should be very quick to do in xml when required. Now when creating a menu we create/load/get the TMenu assets and use that in the TMenu class.

Added colour manipulation to TAM files.

Added a Type ref to the TBinary class as a hint as to what type of data is stored. This value is not USED at all, but will be used as a hint when importing/exporting the TBinary to XML. (All .Asset files will soon be automatticcly exported to a TAX -TootleAssetXml- type so we can manually read the contents or *really* specifiy the Asset data in XML in a very raw format)

Added a TStringLowercase class which always stores the string in lowercase. This is now used in the XML classes so when we look for XMLNodes and properties we don't have to be case-sensitive about it :) We can use this in other places where we need long strings (too long for ref's) but don't need the case (e.g. won't be displaying to the user)

Added a DeviceTypeRef to TDevice's so I can enumerate through them to find specific types.

Fixed the timestamp error when writing a file to a filesys. (was evident when exporting file to a .Asset)

Started on a TAssetBrowser class which creats a render target to show all the assets we have. Not finished yet.

Added a TRenderNodeTile which (kinda hacky atm) repeats a node when rendering to cover the screen on the Z plane, like an infinate background.

Added a "wide" TScreen type. Only tested on PC which just initialises the window to a landscape mode. On the iPod it *will* set everything up to be on it's side with the correct axis' etc