Updated the screenshot handling code and email code.
Unfortunately, we cannot access the filename of the image in  the photo gallery so to remedy this it will now save a jpg file to the temporary directory.  The path of this file will be used to send in an email once this is confirmed to be working as expected.