Updated the core TApplication class with a more complete implementation.  There is still some work to be done on the default application modes.
On the game side the main application manager should overload the TApplication class. When overloading this class you must overload the following pure virtual routine which is used as part of the filesystem creation at initialisation:

virtual TTempString			GetName()	const = 0;	// Name of application - must be specified

You should return a string the which is the same as the product name on the iPod.

There are a variety of virtual routines that may also be overloaded to customise thie classes functionality - see the TLCore::TApplication class definition for further info.

Updated the TLGame::TGame class with a more complete implementation.  Still some work to do with regards to any sub game-modes and Initialisation and shutdown being assumed to occur in a frame. 

Added a TLCore::QuiteRef global TRef to ensure correct TRef checking for this reference.

Added a Bool HasMode(TRefRef) call to the TStateMachine class to be able to check if a mode exists