Implemented first stage of the multi touch support.
Enabled mutitouch processing flag to the UIView to allow mutiple touch events to be passed to our application.

Updated the low level touch event processing routines to  cycle through all touch events rather than just use the first one.

The TTouchData class now stores a few more details such as time stamp and tap count although at this stage these are not used.

Optimised the input system array of the TTouchData to be a data array, TArray, rather than a TPtrArray.

The calls to ForceUpdate() for the input system when a new touch event is processed has been removed as this will likely cause issues with performance more than it currently helps.  All input will now ONLY be updated via the input system so if lag now occurs it is down to us not meeting a reasonable frame rate.

Currently the input still uses the touch data for a single object and I suspect on the iPod itself using mutiple fingers on screen will likely crash or cause visual issues.  I have checked to make sure the input still works on SWTD and Heartist on the simulator that still work as expected.

TODO:
* Add an array of presistent touch objects so that the correct index can be used when processing the input so that each finger touch is uniquely handled.