Implemented final correct version of the debug break for the ipod.
Now the code conditionally compiles in the Debugger() call if it is not a TARGET_OS_IPHONE or is a TARGET_IPHONE_SIMULATOR build.  

All future projects will need to conditionally link to the CoreServices framework in the Debug and Release configurations only for the builds to link correctly when building for the simulator.