Added the call to get the initial preload files for the application.  The application will now load the files during the bootup mode and not carry on to the next mode until the files have been loaded.
You can specify the files that need loading at boot up by overloading the GetPreloadFiles(TArray<TRef>& PreloadFiles) of the TApplication class.

Altered the GetGameObject() call on the TApplication class to be templated.

Changed the debug break when removing the physics or render nodes by the object node to a print.