Implemented initial work on moving over to using the buffered input data - WIP
Updated the sensors to now have multiple label references.  This will provide more flexibility for how to map actions to buttons/axes/POV's etc.

Altered the action map routines by removing the MapActionLabel() and updating the MapAction() routine to use the ref as a label rather than an index.  Essentially all MapActionLabel() calls will need to be renamed to MapAction() and the last ref parameter needs to be a string based id rather than an integer.  All projects have already had these changes made and will be updated via separate commits.

Added default references for buttons.  All buttons will now be assigned a "BN#" label where hash is the index number of the button for the device in question.  This matches the axis and POV's that already use default labels.

Implemented additional keyboard button references.  All keyboard buttons will now have an additional keyboard friendly reference added such as "K_A" for the 'a' key or "K_UP" for the 'up arrow' key.  These generally match the DirectX ID's but mean you do not need to use the special DirectX DIK_### codes and will be usable across platforms.  To map an action to the escape key for example you can now map the action to "BN1" or "K_ESC".

Implemented additional gamepad specific button references.  When a gamepad is created it now checks to see what gamepad it is and will setup some additional pad-specific button references.  At the moment only the Xbox 360 pad is fully supported but additional pads can be easily added. For the 360 pad the 'a' button for example will be assigned a label of "A".  This means you will now be able to map an action to either "BN1" or "A" for this particular pad.  For a PS2 pad you will be able to map an action to "BN1" or "X".  So this will allow use of a generic pad button reference as well as a specific one when/if it's required.