Speed up (Mostly render code)
Whilst trying to work out why something I added was unexpectedly killing the frame rate, I found a bunch of things that probably shouldn't be happening in the render system.
There is a lot less invalidation and re-calculation of bounding boxes now. Local bounds are rarely changed now, and world invalidations are a lot less.

Added some minor flags we can set on nodes now to skip culling tests (e.g. for the world root node) but still enable culling for it's children.

Added sorting to quad tree zones for a bit of a speed up.

Debug now runs at 40fps (fraps limiting) and CPU usage is down to like 5%! Will probably run happily on your laptop now duane :)

Left a comment in EnumDevices to identify where the input system is a massive drag on the update.
Pre-allocated binary data in the keyboard update for a big speed up (a big chunk of the input update anyway).

Added a DebugMesh TRenderNode, which contains it's own mesh for you to freely manipulate as you feel free. Just to make it a lot easier to draw debug lines and boxes and stuff