MAJOR CHANGE: Changed the input system to use buffered data rather than immediate in DirectX.  All immediate code has been commented out and will be removed assuming no issues occur using the buffered mode instead.
Optimised the input buffer by changing it to an array of TInputData (TRef and float) rather than a TBinaryTree.

Changed the device check to be every 10 seconds rather than 3.  Will remove this if it is still a bottle neck on the PC.

Made a few optimisations to various input system calls with TRef where they can be TRefRef's instead