Fiddled with a load of graph functions. Replaced some old slow code with faster versions. I've renamed "UpdateQueue" to "RequestQueue" too, I don't think "UpdateQueue" was a good name really... sounds more like a queue of updates, rather than graph changes.
Fixed a couple of bugs (wrong checks, but don't seem to have come up yet) in the add and remove functions.
Added separate functions for IsInAddQueue and IsInRemoveQueue, and some params to skip checking these queues when looking for nodes to remove redundant array checks.

All this so I could tell if a node was being added or removed :)