Fixed use of "Translate" when it should be "Scale" in the SceneNode_Transform "Transform" message handling. Renamed uses of "Rotate" in the "Transform" messages to "Rotation" to match the TRenderNode usage.
Inserted "SNRef" and "SNType" data into the Message when a transform message is passed on. This will obviously affect the message for anyone else receiving it. If this is a problem we'll need to copy the TMessage first into a new TMessage and send that with the extra data instead.

Corrected some line endings.