Added SetPitch and SetVolume wrapper routines to the audio graph
Updated some uses of the TRef "Pos" to be "Translate" for consistency.

Added a call to the DestroyCharLookupTable() routine in the TRefManager destructor to cleanup the g_RefCharLookup key array correctly on closing the application.

The sphere physics now broadcasts it's render transform.  The render transform will need removing at some stage if possible.

Added preliminary work on making the AddNode() and RemoveNode() graph routine susing a TPtr to be private.  On hold for now whilst other more pressing issues are sorted out.

Removed the TScenegraph::CreateInstance() routine as it shouldn't be used anymore.  Use the graph CreateNode() call instead.

Updated the scheme editor node selection handling to use node refs instead of TPtr's