Renamed the TGui class to be TInputInterface which is more representative of what it is/does. Also renamed the files to correspond with this change.
Altered the input interface code that subscribes to the buttons to be more efficient - gets the button counts and uses the input system to generate the button ref to use rather than cycling through the sensors and checking for sensor type.  Also removed the exposed sensor list as it is now not needed.

Removed public access to the MapAction using a TPtr - all public access should be done via TRef's.