Removed the ChannelID from the TMessage class.  The message ref will now double as a channel ref and message ID. Changed the sender ref to be a member of the TMessage class.  Currently we can create a message without a sender ID but if required we can force the use of a sender ID by removing the single parameter constructor version of the TMessage.
Reverted the physics, render and audio nodes to subscribe directly to the 'owner' scene node.  This is fairly straight forward to re-introduce if the number of redundant messages being sent causes bottlenecks and will simply mean having to create event channels for the required message ID's that are to be sent form a node.

Re-did the conflict/compile issue with the add and remove node notification.  The sender ID is now used for the graph reference instead which should be slightly quicker now it is a member of the TMessage class.

Moved the Quit message handling from project App classes into the base TApplication class.

Removed teh SNRef and SNType extra data from being attached to the transofrm node messages.  The SNType wasn't being used and the SNRef is essentially the sender ref so use that instead.

Went through and renamed lots of messages to use more appropriate TRef values to represent what they do rather than what is creating them or any kind of group name.