Added basic window activation/deactivation support that will disable the input system when the game window receives a WM_ACTIVATE message. We may need to extend this at some stage to pause the game rather than just disabling the input but for now should resolve any issues of the input processing whilst the window is inactive.
FYI - The GWinControl class is now a publisher so we can potentially broadcast other windows events and handle them in game such as window resizes etc. The TScreen subscribes to the window control and in turn the screen manager subscribes to the TScreens to be able to pass the messages onto other things like the input system.  There shouldn't be much of an overhead for this (mostly memory) as not many messages will be sent via this system and I don't expect many things to subscribe to the screen manager in the long term.