Removed the TAudioInterface files that were now redundant.
Updated the audio graph with the interface routines that were in the TAudioInterface. The audio graph should be used when altering/requesting audio outside of the node systems which have some wrapper routines you can use instead.

Added the ability to mute the audio system.  If you have a keyboard attached the 'M' key will toggle the mute of all audio.  We can map this to a menu option simply by sending a 'Mute' action to the audio graph.

Added some low level code to disable the audio system when the app loses focus and enable it again when it becomes active again.  Upon testing this didn't actually do what it said on the tin although from the docs and info I gathered form the internet it should work so I am leaving it in just in case it is just my sound card not fully implementing the OpenAL standard.  NOTE: I have a second card I can test it with as apparently you get can different results with different cards.

Removed code to subscribe the input manager to the screen manager in the TApplication init.  This has now moved into the input manager code instead and done when the event channels are added so we can guarantee the event channel exists.

Updated the object audio node creation to go through the audio graph.

The screenmanager now registers the ScreenChanged and OnWindowChanged event channels correctly.