Added an m_OwnerSceneNode variable to RenderNodes, PhysicsNodes and audio nodes to keep track of the scene node that created us. (This is the same node we get subscribed to)
Changed the OnTransform message handling of SceneNodeTransform to "DoTransform". Again, having this base node type automaticcly apply translations causes bugs. Instead this functionality has been moved to the Object scene node, and we only copy the transform if the sender is our physics node (PhysicsNode changes, we change ourselves, then change render node)
Similarly the render node only changes it's transform if OnTransform comes from it's owner[scene node].

Added Capsule shape, collision shape and various intersection math routines.

Added a TRenderNodePhysicsNode debug render node which draws the collision shape of the assigned phyiscs node and subscribes to changes of this shape.

Corrected an issue with the physics node only sending out transform changes if the movement was too much. Small car movements wouldnt get reflected to the scene node and then to the render node. The render of the car is always a frame or two out, but this made it much more laggy.