Disabled the "faster quad tree zone update" as it's not accurate enough. (was obvious when driving around) Back to brute force method.
Consolidated all transform changes in physics nodes until the PostUpdate. We now only send out a max of 1 "OnTransform" message per update, instead of a message for every type-of-transform change. Big speed up on Cops.

Removed the automatic calculation of the bounding BOX in the render node ZoneNode, this means we only calc a sphere when doing the quad tree intersection tests. Bit of a speed up. This will be improved again with Trac entry #46. We need to balance amount of geometry we render and the quad tree traversal though, so the box check may go back in. (probably when that trac entry is done)
http://grahamgrahamreeves.getmyip.com:1984/Trac/ticket/46