Trac ticket #46: Moved the latest scene transform (WorldTransform) onto the render node on every render. Now world position/bounds/datums are calculated when requested now, not at render time. The system now marks world shapes validity as syncbools; SyncFalse when invalid, SyncWait when old (newer scenetransform is available or world has been invalidated by parent or our own local transform changes) and SyncTrue when up to date. The same applies for the world transform itself.
This all means that world bounds/positions are all still valid even when old. It also means we no longer need the Calc_XXX flags on the render node as we can calculate any datum/bounds at any time (as long as we have been rendered at least once). This also means we can ditch the flag setting on the Widgets which simplifies that and makes it much easier to click on any render node at all. 
This gives us a minor speed boost as we don't recalculate things as much as we did before, and/or don't calculate bounds that we don't use. We can also catch very small transform changes (maybe even non-changes) and ignore them and restore old transformed datums/shapes, which would be a BIG saving.
The main benefit is flexibility and means we can attach datums to render nodes (or extract them out of the mesh and request them at runtime) much much more easliy.