Added a CreateValidTRef(s32 value) routine to the TLRef namespace.  This essentially will mask the value passed in with the g_invalidrefmask and return a value which will always be valid.  This should only really be used when using memory addresses or random numbers as a unique ID for a TRef where they could potentally be invalid.
Enabled the multi-touch handling for the iPod.  Completed first stage of updating the multi-touch handling as detailed below.

Updated the multi-touch on the ipod to work as expected at a low level.  The touch events are now mapped correctly to a corresponding game object allowing us to track button 1 as the first finger down, button 2 as the second and so on.  The system is limited to 4 fingers and in debug more than 4 fingers will trigger a debug break.

Removed the old ipod input processing code that I commented out a while ago as this is now redundant.

Made a handful of input accessor inline routines FORCEINLINE instead

Updated the sensors to now have a cursor index associated with them.  On the PC this will always be cursor 0 (as there is only one anyway) but on the ipod button 0 will use cursor 0, button 1 will use cursor 1 and the same for the axes.  This corrects issues with the test app where pressing a button would activate the current button a user had selected because the first cursor was always being used on the ipod.