Updated the TInputInterface to not send out a message prematurely when a button is released.  This semi-fixes the cops gui issues.  Now when you press both accelerate or decelerate buttons they will work independently as expected.
However there is still an issue for when you press the button in the GUi with the mouse and also use the keyboard button for the same action where when you release either one it thinks you have stopped that action when you haven't.  I don't believe this issue is to do with the TInputInterface though as I think it is down to having two gui controllers - the default one which uses the mouse and one for the keyboard - but I will investigate this further at a later stage.

I am committing this so I can test it still works OK on the iPod mainly. :)