Ditched the TFrustum class used on the camera... may delete this shape type entirely from the maths... Instead the [projection] camera just stores a 2D box and an oblong box for frustum tests.
Added rendering of the 2D frustum box to the rendertarget's debug render. The Green box is the Oblong view frustum (so is always correctly positioned/sized). The new red box is the 2D AA box of the frustum. This is the box we do all our visibility culling and zone testing with for the camera. Should be a bit more obvious that things are/are not being culled correctly during render.

Removed the parent-in-request-queues code from the graphs. Now blindly adds a node to the request queue, and upon add if the parent isn't in the graph any more it doesn't complain, it just doesn't get put in the graph. Added a define to swap this around so on add/createnode it does the parent-will-be-removed-from-graph test. Newer version is faster.

Added a define to enable the "Node removed/removing from graph" messages. Graph now only prints out stuff that is unexpected. Correct-behaviour has no prints.

Consolidate the way render nodes calculate bounds boxes to a template TBoundsShape type. This is to aid the future adding of datum access, and has allowed me to add more shapes quicker (i.e. 2D bounds). Another shape type to conslidate with the maths/physics shapes, although this has some render-node-specific stuff. (Although that could help the physics too). 

Now we have 2D bounds types on render nodes these are used a bit more for the zone tests, widget click tests etc and are faster.

Added "params" for quad trees to dictate certain behaviours which can be customised for different cases to get optimum use of the trees for different purposes. In it's own little class so I can easily add more params later without fiddling with constructors etc.

Added a QuadTree for the scene graph. This will be controlling the Sleep/Awake state of scene nodes that are too far away and don't want updating.

Changed TSceneNode_Transform::OnTransformChanged to only send out validate transform data.