Added TLDebug_Assert macro.
Removed TMatrix from TTransform for a bit of a speed boost and because we never use it. Simplifies the structure and means we don't have to implement a load of things that were missing. Changed the IsValid bools to a bit mask. Added an assert if you access the const translate/scale/rotation (which you must be reading as its const) when it's not valid.

Added IsZero and IsNonZero function tests to the TypeX types. 

Added a check for IsNonZero in some of the TTransform code as the rotation/scale has no effect if the vector passed in is zero.

Changed the GetPosition of the base scene transform to just return the translate in the transform as 0,0,0 will never be anything other than the translate.

Derived SceneNode_Transform from a TQuadTreeNode for the future enabling of movement around the scenegraph zones.

Moved the camera matrix onto it's own matrix in the TCamera. This was the only usage of the Matrix in the TTransform we had. Also removed the redundant cache's of the current modelview & projection matrixies of the camera/render target as we never need them.

Not checked the IPod that it runs, but it should compile.