BIG CHANGE! The TPtr system didn't work very well with multiple inheritance. I've sorted that by ditching the intermediate TCounter class. Instead, the TPtr itself stores a pointer to the object, in it's native type. This keep all the RTTI and this-offset from the memory. Before the templated CounterType would lose this information as it was cast. Now, a u32 counter is shared across the TPtrs. We ASSUME the object pointer is valid, the assumption being, if we still have a counter then the object must be valid. (it makes sense :)
Now, we can cast TPtrs around to the different classes in multiple inherited classes without corrupting this pointers. The static_casts now work properly too. 

This system should probably be a good bit faster too! less memory jumps/cache misses.

Now I can, the TSceneNode_Transform object derives from a TQuadTreeNode as well as the base SceneNode type. The PhysicsNode can be done in the same way now (and can ditch the TPhysicsZoneNode) class.
Added (currently unused) active-zone variable to define what zone is active in the scene graph. We will use this zone to determine what objects to update. (Any nodes in this zone and neighbour zones and all non-sleep nodes)

Added an Active state to the quad tree zones. It's a bit of a pain to derive Zone class, and I'm hoping this is all I need... Added appropriate functions to the SceneNode_Transform class for OnWake (zone was inactive, now active) and OnSleep (zone was active, now inactive).
Changed the RenderNodeQuadTree to render the zone spheres as outlines if the zone is inactive.