Added a IsEnabled() function to all graph node types. This is NOT virtual though, just needs implementing on any new graph node types. Rendernode and phsyics node check against their flag (new flag for physics). Audio and scene nodes always return TRUE. For this I've added a This() function on the TGraphNode class which safely casts to its T type. This is just to enable more templated function calling without the need for virtuals (compile time resolution so bit of speed up, less vtable entries, no virtual function overhead etc). Just use like This().Function(), This().m_Variable etc.
Added a OnZoneChanged callback when a QuadTreeNode moves to a different zone.

Now the TPtr has been fixed for multiple inheritance, I've deleted the PhysicsZoneNode class and moved all the QuadTreeNode virtuals into the physics node (faster, better etc) and made all the relavant changes to the physics graph code.

Added a missing GraphNode::Update to the physics node update. Physics AND scene graph now overload the UpdateAll() function. PostUpdate (called right after the update, and before next node's update) has been added to the scene node. This is mostly for the zone code really to update the zone. Because of the Update ordering for derived types it would put the code in the wrong place when the base Update was called.

Scenegraph now updates nodes via the active zone (if a root zone is set) so only nodes in an active zone (and soon, it's neighbour zones) will get updated.

Scene nodes do their In-Zone tests now based on their position (NOT SHAPE). This means a node should always be in a leaf zone. To facilitate this, I may switch the scenegraph system to be a grid rather than a tree as there's not much benefit in it being a tree...

Added OnZoneSleep and OnZoneWake functions to the SceneNode_Transform class to catch when the node will no longer/will now be updated.
The SceneNode_Object class overloads this to Enable/Disable its render and physics nodes. When CreateRenderNode and CreatePhysicsNode is called, it checks it's current zone and if it's not active they have their respective Enable flags turned off for their Initialise. More stuff like this coming soon.