Added a TArray Copy() and Add() to copy/append an array of a different type. Should work if TYPE=OTHERTYPE works.
Added a QuadTreeZone func to find a the zone a specific point is in. Will expand at some point to use Shapes.

Added functionality to the scenenode_object code so when transform(or trans/rot/scale etc) is explicitly set with SetTransform/SetTranslate etc the physics node's transform is also explicitly set, and the physics node's changes are not published. This is for when you want to move your node around and so the physics won't just overwrite your change again next frame.