Lowered resolution of TMesh spheres and disabled the Geometry checking for now. (too slow)
Added a TPathLink list in network paths. These are simple Ref/Ref links with a pre-calculated line and bounding sphere. Very useful when looking for "roads" and not just nodes in the path.

Added an AddNew function to the key array. Add's a new Pair entry with the appropriate key and returns the Item. It was just handy for a certain situation :)

Added a function to get a eular 2D angle from a quaternion. Works off of the Z axis at the moment so only really appropriate for a few things. can be expanded later. is probably not the most effecient way of calculating it either but it's accurate.

Added a class to the game lib for caching PathLink/Zone intersection information. It essentially stores a list of all the network pathlinks that are intersecting a zone. And all the zones a path link is in.

Fixed an access to the render node glyph's translate when it wasn't valid.

Expanded the active zones in the scenegraph to be the active zone, and it's neighbours as fully-awake. *Their* neighbours are then half-awake. This means we have more zones active at once (25) but it does mean if we're at the edge of the active zone the zone right next to us will be active too, this gets rid of popping issues, lack of collision etc etc.
The whole list of active/half active zones is now kept as this is faster to update too.