Added a TAtlas asset type. This is a texture atlas for a texture. (Provides UV mapping information for a "glyph" contained inside the atlas). This is used currently (but not exclusively) for textured fonts.
Added support for .fnt files generated by the program "Bitmap font generator". Very good, very simple system that makes up a PNG (with appropriate alphas) and an XML "layout" of the characters inside the texture (an atlas).
A characters height is on a scale of 0-1. (Truetype fonts are supposed to be this scale too but need adjusting it seems)

Added support for a textured-font text render node. The former Text render node is now a "vector" render node, both derive from a font-rendernode type, which will deal with text layout/alignment etc at a base level. This textured font render node combines a texture atlas and a texture into a render node which generates a mesh of triangles to draw quads of characters.

Added a RemoveNode function to the graph which takes a non-const reference to a render node and will check if it's valid before trying to remove the node, and then invalidate it afterwards. This is just to simplify the removal of a render node. If you have a member you need to delete just call
Graph->RemoveNode( m_RenderNode );
no need for IsValid() checks beforehand and SetInvalid() afterwards any more.