Implemented first pass asset script (timeline) handling. We can now scale and translate nodes via a simple keyframe style XML script.  Rotations are still WIP and are disabled atm.
Rewrote the asset script XML and .asset import and export code.  The correct number of keys, nodes and commands are now created from the .asset file.

Implemented the TAssetScriptInstance class.  This class is used as an instance of an asset script, keeping it's own time through the asset script / timeline.  At the moment it is necessary to call the Update() routine manually with the timestep but at some point I will add these to the asset script manager so that we can have auto-updating asset scripts.

You specify the file asset to use when instancing the class such as:
   TPtr<TLAnimation::TAssetScriptInstance> pTimeline = new TLAnimation::TAssetScriptInstance("testanim").

You can bind an asset script to a node ref and graph ref by calling BindTo() on a newly created TAssetScriptInstance.  
  pTimeline->BindTo("copcar1");
When a script command is processed if the script command uses a special "this" reference for the Node ref then it will replace this reference with the node reference specified in the BindTo() call.  This means you can have an asset script control a node instance directly rather than relying on the ref being correct in the asset script XML file, allowing for one script file to be used for many node instacnes.

Added "Rotate", "Translate" and "Scale" message refs to the Render node.  This is because the commands are independant and can be sent as-is from the command processing so no intermediate message with the OnTransform ref is used.  This can be changed if required but will probably need us to go through exactly how we want the commands to work from a messaging point of view.

For examples of current asset script XML usage check the FindFritz animtest and tllimb xml files.